requirements.txt
streamlit
google-generativeai
pip install -r requirements.txt